refactor(storage): Keep the search capabilities in capabilities.go
SearchCapabilities sat in reader.go among the query and result types, while the FilterCapabilities it nests sat in filter.go. Both now live in capabilities.go, search-wide first and the nested one after it, with their tests together in capabilities_test.go. reader_test.go held only the capability tripwire, so it goes with it.
